Remove Options <-> Parser cycle.
ScopedFlagNameForFuzzyMatching just needed to be moved from Options to
Parser to break the cycle. With this done, restore full typing
information for FlagNameProvider.

# Rust tests will be skipped. Delete if not intended.
[ci skip-rust-tests]
